+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file. Starting wi
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+  * (quick hack) Optional trailing parameter `order` on SequelizeBaseService.getAssociatedMany()
+### Removed
+### Changed
+### Fixed
+### Deprecated
+### Security
+
+
 ## [3.1.1] -- 2020-01-25
 
 ### Added